refactor: extract project selection to custom hook
